Extract app markdown export formatter
