+++ src/appExportMarkdown.ts
@@ -1,0 +1,73 @@
+import type { UiMessage, UiThread } from './types/codex'
+
+export function buildThreadMarkdown(thread: UiThread | null, messages: UiMessage[]): string {
+  const lines: string[] = []
+  const threadTitle = thread?.title?.trim() || 'Untitled thread'
+  lines.push(`# ${escapeMarkdownText(threadTitle)}`)
+  lines.push('')
+  lines.push(`- Exported: ${new Date().toISOString()}`)
+  lines.push(`- Thread ID: ${thread?.id ?? ''}`)
+  lines.push('')
+  lines.push('---')
+  lines.push('')
+
+  for (const message of messages) {
+    const roleLabel = message.role ? message.role.toUpperCase() : 'MESSAGE'
+    lines.push(`## ${roleLabel}`)
+    lines.push('')
+
+    const normalizedText = message.text.trim()
+    if (normalizedText) {
+      lines.push(normalizedText)
+      lines.push('')
+    }
+
+    if (message.commandExecution) {
+      lines.push('```text')
+      lines.push(`command: ${message.commandExecution.command}`)
+      lines.push(`status: ${message.commandExecution.status}`)
+      if (message.commandExecution.cwd) {
+        lines.push(`cwd: ${message.commandExecution.cwd}`)
+      }
+      if (message.commandExecution.exitCode !== null) {
+        lines.push(`exitCode: ${message.commandExecution.exitCode}`)
+      }
+      lines.push(message.commandExecution.aggregatedOutput || '(no output)')
+      lines.push('```')
+      lines.push('')
+    }
+
+    if (message.fileAttachments && message.fileAttachments.length > 0) {
+      lines.push('Attachments:')
+      for (const attachment of message.fileAttachments) {
+        lines.push(`- ${attachment.path}`)
+      }
+      lines.push('')
+    }
+
+    if (message.images && message.images.length > 0) {
+      lines.push('Images:')
+      for (const imageUrl of message.images) {
+        lines.push(`- ${imageUrl}`)
+      }
+      lines.push('')
+    }
+  }
+
+  return `${lines.join('\n').trimEnd()}\n`
+}
+
+export function buildExportFileName(thread: UiThread | null): string {
+  const threadTitle = thread?.title?.trim() || 'chat'
+  const sanitized = threadTitle
+    .toLowerCase()
+    .replace(/[^a-z0-9]+/g, '-')
+    .replace(/^-+|-+$/g, '')
+  const base = sanitized || 'chat'
+  const stamp = new Date().toISOString().replace(/[:.]/g, '-')
+  return `${base}-${stamp}.md`
+}
+
+export function escapeMarkdownText(value: string): string {
+  return value.replace(/([\\`*_{}\[\]()#+\-.!])/g, '\\$1')
+}
